+++ pyelasticsearch/__init__.py
@@ -3,15 +3,14 @@ from __future__ import absolute_import
 from pyelasticsearch.client import ElasticSearch
 from pyelasticsearch.exceptions import (Timeout, ConnectionError,
                                         ElasticHttpError,
-                                        InvalidJsonResponseError,
                                         ElasticHttpNotFoundError,
-                                        IndexAlreadyExistsError)
+                                        IndexAlreadyExistsError,
+                                        InvalidJsonResponseError)
 
-__author__ = 'Robert Eanes'
-__all__ = ['ElasticSearch', 'ElasticHttpError', 'InvalidJsonResponseError',
-           'Timeout', 'ConnectionError', 'ElasticHttpNotFoundError',
-           'IndexAlreadyExistsError']
-__version__ = '0.6.1'
+__author__ = 'Erik Rose'
+__all__ = ['ElasticSearch', 'ElasticHttpError', 'Timeout', 'ConnectionError',
+           'ElasticHttpNotFoundError', 'IndexAlreadyExistsError']
+__version__ = '1.0'
 __version_info__ = tuple(__version__.split('.'))
 
 get_version = lambda: __version_info__