+++ pyelasticsearch/exceptions.py
@@ -1,4 +1,4 @@
-from requests import Timeout, ConnectionError
+from urllib3.exceptions import TimeoutError as Timeout, ConnectionError
 
 
 class ElasticHttpError(Exception):
@@ -40,8 +40,9 @@ class InvalidJsonResponseError(Exception):
     Exception raised in the unlikely event that ES returns a non-JSON response
     """
     @property
-    def response(self):
+    def input(self):
+        """Return the data we attempted to convert to JSON."""
         return self.args[0]
 
     def __unicode__(self):
-        return u'Invalid JSON returned from ES: %r' % (self.response,)
+        return u'Invalid JSON returned from ES: %r' % (self.input,)