+++ pyelasticsearch/tests/client_tests.py
@@ -1,11 +1,7 @@
 # coding=utf-8
-import sys
-import unittest
 
-from mock import patch
+from mock import ANY, patch
 from nose.tools import eq_, ok_, assert_raises, assert_not_equal
-import requests
-import six
 
 # Test that __all__ is sufficient:
 from pyelasticsearch import *
@@ -22,18 +18,18 @@ class IndexingTestCase(ElasticSearchTestCase):
 
     def test_indexing_with_id(self):
         result = self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=1)
-        self.assert_result_contains(result, {'_type': 'test-type', '_id': '1', 'ok': True, '_index': 'test-index'})
+        self.assert_result_contains(result, {'_type': 'test-type', '_id': '1', '_index': 'test-index'})
 
     def test_indexing_with_0_id(self):
         result = self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=0)
-        self.assert_result_contains(result, {'_type': 'test-type', '_id': '0', 'ok': True, '_index': 'test-index'})
+        self.assert_result_contains(result, {'_type': 'test-type', '_id': '0', '_index': 'test-index'})
 
     def test_indexing_with_unicode(self):
         """Test unicode field values and path components."""
         unicode_name = u'Jöe Téster'
         unicode_id = u'smöö'
         result = self.conn.index('test-index', 'test-type', {'name': unicode_name}, id=unicode_id)
-        self.assert_result_contains(result, {'_type': 'test-type', '_id': unicode_id, 'ok': True, '_index': 'test-index'})
+        self.assert_result_contains(result, {'_type': 'test-type', '_id': unicode_id, '_index': 'test-index'})
 
         # Make sure it comes back out intact:
         result = self.conn.get('test-index', 'test-type', unicode_id)
@@ -44,13 +40,13 @@ class IndexingTestCase(ElasticSearchTestCase):
 
     def test_quoted_chars_in_id(self):
         result = self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id="""<>?,./`~!@#$%^&*()_+=[]\{{}|:";'""")
-        self.assert_result_contains(result, {'_type': 'test-type', '_id': """<>?,./`~!@#$%^&*()_+=[]\{{}|:";'""", 'ok': True, '_index': 'test-index'})
+        self.assert_result_contains(result, {'_type': 'test-type', '_id': """<>?,./`~!@#$%^&*()_+=[]\{{}|:";'""", '_index': 'test-index'})
 
     def test_indexing_without_id(self):
         result = self.conn.index(
             'test-index', 'test-type', {'name': 'Joe Tester'})
         self.assert_result_contains(result,
-            {'_type': 'test-type', 'ok': True, '_index': 'test-index'})
+            {'_type': 'test-type', '_index': 'test-index'})
         # should have an id of some value assigned.
         ok_('_id' in result and result['_id'])
         # should not generate the same id twice
@@ -60,20 +56,24 @@ class IndexingTestCase(ElasticSearchTestCase):
 
     def test_explicit_index_create(self):
         result = self.conn.create_index('test-index')
-        self.assert_result_contains(result, {'acknowledged': True, 'ok': True})
+        self.assert_result_contains(result, {'acknowledged': True})
 
     def test_close_index(self):
         """Make sure a close_index call on an open index reports success."""
         self.conn.create_index('test-index')
+        self.conn.health('text-index', wait_for_status='yellow', timeout=1)
+
         result = self.conn.close_index('test-index')
-        self.assert_result_contains(result, {'acknowledged': True, 'ok': True})
+        self.assert_result_contains(result, {'acknowledged': True})
 
     def test_open_index(self):
         """Make sure an open_index call on a closed index reports success."""
         self.conn.create_index('test-index')
+        self.conn.health('text-index', wait_for_status='yellow', timeout=1)
         self.conn.close_index('test-index')
+
         result = self.conn.open_index('test-index')
-        self.assert_result_contains(result, {'acknowledged': True, 'ok': True})
+        self.assert_result_contains(result, {'acknowledged': True})
 
     def test_get_settings(self):
         self.conn.create_index('test-index')
@@ -109,21 +109,23 @@ class IndexingTestCase(ElasticSearchTestCase):
             'GET', ['_cluster', 'health', ''], query_params={})
 
     def test_cluster_state(self):
-        result = self.conn.cluster_state(filter_routing_table=True)
-        ok_('nodes' in result)
-        self.assertFalse('routing_table' in result)
+        """Make sure "_all" works for an index specifier and metric filtering
+        works."""
+        result = self.conn.cluster_state(metric='routing_table')
+        ok_('routing_table' in result)
+        ok_('master_node' not in result)
 
     def test_delete_by_id(self):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=1)
         self.conn.refresh(['test-index'])
         result = self.conn.delete('test-index', 'test-type', 1)
-        self.assert_result_contains(result, {'_type': 'test-type', '_id': '1', 'ok': True, '_index': 'test-index'})
+        self.assert_result_contains(result, {'_type': 'test-type', '_id': '1', '_index': 'test-index'})
 
     def test_delete_by_0_id(self):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=0)
         self.conn.refresh(['test-index'])
         result = self.conn.delete('test-index', 'test-type', 0)
-        self.assert_result_contains(result, {'_type': 'test-type', '_id': '0', 'ok': True, '_index': 'test-index'})
+        self.assert_result_contains(result, {'_type': 'test-type', '_id': '0', '_index': 'test-index'})
 
     def test_delete_by_id_without_id(self):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=1)
@@ -137,29 +139,30 @@ class IndexingTestCase(ElasticSearchTestCase):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=1)
         self.conn.refresh(["test-index"])
         result = self.conn.delete_all("test-index", "test-type")
-        self.assert_result_contains(result, {'ok': True})
+        self.assert_result_contains(result, {u'acknowledged': True})
 
     def test_delete_by_query(self):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Tester'}, id=1)
         self.conn.index('test-index', 'test-type', {'name': 'Bill Baloney'}, id=2)
         self.conn.index('test-index', 'test-type', {'name': 'Horace Humdinger'}, id=3)
-        self.conn.refresh(['test-index'])
 
         self.conn.refresh(['test-index'])
         result = self.conn.count('*:*', index=['test-index'])
         self.assert_result_contains(result, {'count': 3})
 
         result = self.conn.delete_by_query('test-index', 'test-type', {'query_string': {'query': 'name:joe OR name:bill'}})
-        self.assert_result_contains(result, {'ok': True})
+
+        eq_(result['_indices']['test-index']['_shards']['failed'], 0)
 
         self.conn.refresh(['test-index'])
         result = self.conn.count('*:*', index=['test-index'])
+        # Only Horace should be left:
         self.assert_result_contains(result, {'count': 1})
 
     def test_delete_index(self):
         self.conn.create_index('another-index')
         result = self.conn.delete_index('another-index')
-        self.assert_result_contains(result, {'acknowledged': True, 'ok': True})
+        self.assert_result_contains(result, {'acknowledged': True})
 
     def test_delete_nonexistent_index(self):
         """
@@ -179,7 +182,7 @@ class IndexingTestCase(ElasticSearchTestCase):
     def test_put_mapping(self):
         result = self.conn.create_index('test-index')
         result = self.conn.put_mapping('test-index', 'test-type', {'test-type': {'properties': {'name': {'type': 'string', 'store': 'yes'}}}})
-        self.assert_result_contains(result, {'acknowledged': True, 'ok': True})
+        self.assert_result_contains(result, {'acknowledged': True})
 
     def test_get_mapping(self):
         result = self.conn.create_index('test-index')
@@ -187,32 +190,32 @@ class IndexingTestCase(ElasticSearchTestCase):
         self.conn.put_mapping('test-index', 'test-type', mapping)
 
         result = self.conn.get_mapping(index=['test-index'], doc_type=['test-type'])
-        eq_(result, {'test-type': {'properties': {'name': {'type': 'string', 'store': True}}}})
+        eq_(result, {u'test-index': {u'mappings': {u'test-type': {u'properties': {u'name': {u'type': u'string', u'store': True}}}}}})
 
     def test_index_status(self):
         self.conn.create_index('another-index')
         result = self.conn.status('another-index')
         self.conn.delete_index('another-index')
         ok_('indices' in result)
-        self.assert_result_contains(result, {'ok': True})
+        eq_(result['_shards']['failed'], 0)
 
     def test_index_flush(self):
         self.conn.create_index('another-index')
         result = self.conn.flush('another-index')
         self.conn.delete_index('another-index')
-        self.assert_result_contains(result, {'ok': True})
+        eq_(result['_shards']['failed'], 0)
 
     def test_index_refresh(self):
         self.conn.create_index('another-index')
         result = self.conn.refresh('another-index')
         self.conn.delete_index('another-index')
-        self.assert_result_contains(result, {'ok': True})
+        eq_(result['_shards']['failed'], 0)
 
     def test_index_optimize(self):
         self.conn.create_index('another-index')
         result = self.conn.optimize('another-index')
         self.conn.delete_index('another-index')
-        self.assert_result_contains(result, {'ok': True})
+        eq_(result['_shards']['failed'], 0)
 
     def test_bulk_index(self):
         # Try counting the docs in a nonexistent index:
@@ -224,8 +227,8 @@ class IndexingTestCase(ElasticSearchTestCase):
         ]
         result = self.conn.bulk_index('test-index', 'test-type', docs)
         eq_(len(result['items']), 2)
-        eq_(result['items'][0]['create']['ok'], True)
-        eq_(result['items'][1]['index']['ok'], True)
+        eq_(result['items'][0]['create']['status'], 201)
+        eq_(result['items'][1]['index']['status'], 201)
         eq_(result['items'][1]['index']['_id'], '303')
         self.conn.refresh()
         eq_(self.conn.count('*:*',
@@ -233,13 +236,8 @@ class IndexingTestCase(ElasticSearchTestCase):
 
     def test_error_handling(self):
         # Wrong port.
-        conn = ElasticSearch('http://localhost:1009200/')
+        conn = ElasticSearch('http://localhost:3355/')
         assert_raises(ConnectionError, conn.count, '*:*')
-
-        # Test invalid JSON.
-        resp = requests.Response()
-        resp._content = six.b('{"busted" "json" "that": ["is] " wrong')
-        assert_raises(InvalidJsonResponseError, conn._decode_response, resp)
 
     def test_update(self):
         """Smoke-test the ``update()`` API."""
@@ -266,26 +264,18 @@ class IndexingTestCase(ElasticSearchTestCase):
             ]
         }
         result = self.conn.update_aliases(settings)
-        self.assert_result_contains(result, {'acknowledged': True, 'ok': True})
+        self.assert_result_contains(result, {'acknowledged': True})
 
     def test_alias_nonexistent_index(self):
-        settings = {
-            "actions": [
-                {"add": {"index": "test1", "alias": "alias1"}}
-            ]
-        }
+        actions = [{"add": {"index": "test1", "alias": "alias1"}}]
         assert_raises(ElasticHttpNotFoundError,
-                          self.conn.update_aliases,
-                          settings)
+                      self.conn.update_aliases,
+                      actions)
 
     def test_list_aliases(self):
         self.conn.create_index('test-index')
-        settings = {
-            "actions": [
-                {"add": {"index": "test-index", "alias": "test-alias"}}
-            ]
-        }
-        self.conn.update_aliases(settings)
+        actions = [{"add": {"index": "test-index", "alias": "test-alias"}}]
+        self.conn.update_aliases(actions)
         result = self.conn.aliases('test-index')
         eq_(result, {u'test-index': {u'aliases': {u'test-alias': {}}}})
 
@@ -306,30 +296,34 @@ class IndexingTestCase(ElasticSearchTestCase):
 
     def test_percolate(self):
         self.conn.create_index('test-index')
+        self.conn.put_mapping(
+            'test-index',
+            'test-type',
+            {'test-type': {'properties': {'name': {'type': 'string'}}}})
 
         # Index a few queries in the percolator
         result = self.conn.index(
-            '_percolator',
             'test-index',
+            '.percolator',
             {'query': {'match': {'name': 'Joe'}}},
             id='id_1')
         result = self.conn.index(
-            '_percolator',
             'test-index',
+            '.percolator',
             {'query': {'match': {'name': 'not_that_guy'}}},
             id='id_2')
 
         # Percolate a document that should match query ID 1:
         document = {'doc': {'name': 'Joe'}}
-        result = self.conn.percolate('test-index','test-type', document)
-        self.assert_result_contains(result, {'matches': ['id_1'], 'ok': True})
+        result = self.conn.percolate('test-index', 'test-type', document)
+        self.assert_result_contains(
+            result,
+            {'matches': [{'_id': 'id_1', '_index': 'test-index'}]})
 
         # Percolate a document that shouldn't match any queries
         document = { 'doc': {'name': 'blah'} }
         result = self.conn.percolate('test-index', 'test-type', document)
-        self.assert_result_contains(result, {'matches': [], 'ok': True})
-
-        self.conn.delete_index('_percolator')
+        self.assert_result_contains(result, {'matches': []})
 
 
 class SearchTestCase(ElasticSearchTestCase):
@@ -345,15 +339,15 @@ class SearchTestCase(ElasticSearchTestCase):
 
     def test_multi_get_simple(self):
         result = self.conn.multi_get([1], index='test-index', doc_type='test-type')
-        self.assert_result_contains(result, {'docs': [{'_type': 'test-type', '_id': '1', '_source': {'name': 'Joe Tester'}, '_index': 'test-index', "_version": 1, "exists": True}]})
+        self.assert_result_contains(result, {u'docs': [{u'_type': u'test-type', u'_source': {u'name': u'Joe Tester'}, u'_index': u'test-index', u'_version': 1, u'found': True, u'_id': u'1'}]})
 
     def test_multi_get_mix(self):
         result = self.conn.multi_get([{'_type': 'test-type', '_id': 1}], index='test-index')
-        self.assert_result_contains(result, {'docs': [{'_type': 'test-type', '_id': '1', '_source': {'name': 'Joe Tester'}, '_index': 'test-index', "_version": 1, "exists": True}]})
+        self.assert_result_contains(result, {u'docs': [{u'_type': u'test-type', u'_source': {u'name': u'Joe Tester'}, u'_index': u'test-index', u'_version': 1, u'found': True, u'_id': u'1'}]})
 
     def test_multi_get_custom(self):
         result = self.conn.multi_get([{'_type': 'test-type', '_id': 1, 'fields': ['name'], '_index': 'test-index'}])
-        self.assert_result_contains(result, {'docs': [{'_type': 'test-type', '_id': '1', 'fields': {'name': 'Joe Tester'}, '_index': 'test-index', "_version": 1, "exists": True}]})
+        self.assert_result_contains(result, {'docs': [{'_type': 'test-type', '_id': '1', 'fields': {'name': ['Joe Tester']}, '_index': 'test-index', "_version": 1, "found": True}]})
 
     def test_get_count_by_search(self):
         result = self.conn.count('name:joe', index='test-index')
@@ -361,7 +355,7 @@ class SearchTestCase(ElasticSearchTestCase):
 
     def test_search_by_field(self):
         result = self.conn.search('name:joe', index='test-index')
-        self.assert_result_contains(result, {'hits': {'hits': [{'_score': 0.19178301, '_type': 'test-type', '_id': '1', '_source': {'name': 'Joe Tester'}, '_index': 'test-index'}], 'total': 1, 'max_score': 0.19178301}})
+        self.assert_result_contains(result, {'hits': {'hits': [{'_score': ANY, '_type': 'test-type', '_id': '1', '_source': {'name': 'Joe Tester'}, '_index': 'test-index'}], 'total': 1, 'max_score': ANY}})
 
     def test_search_string_paginated(self):
         with patch.object(self.conn, 'send_request') as send_request:
@@ -401,7 +395,7 @@ class SearchTestCase(ElasticSearchTestCase):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Test'}, id=3)
         self.conn.refresh(['test-index'])
         result = self.conn.more_like_this('test-index', 'test-type', 1, ['name'], min_term_freq=1, min_doc_freq=1)
-        self.assert_result_contains(result, {'hits': {'hits': [{'_score': 0.19178301, '_type': 'test-type', '_id': '3', '_source': {'name': 'Joe Test'}, '_index': 'test-index'}], 'total': 1, 'max_score': 0.19178301}})
+        self.assert_result_contains(result, {'hits': {'hits': [{'_score': ANY, '_type': 'test-type', '_id': '3', '_source': {'name': 'Joe Test'}, '_index': 'test-index'}], 'total': 1, 'max_score': ANY}})
 
     def test_mlt_with_body(self):
         self.conn.index('test-index', 'test-type', {'name': 'Joe Test', 'age': 22}, id=2)
@@ -423,7 +417,7 @@ class SearchTestCase(ElasticSearchTestCase):
             }
         result = self.conn.more_like_this('test-index', 'test-type', 1, ['name'], body=body, min_term_freq=1, min_doc_freq=1)
         self.assert_result_contains(result,
-                {'hits': {'hits': [{'_score': 0.19178301, '_type': 'test-type', '_id': '3', '_source': {'age': 16, 'name': 'Joe Justin'}, '_index': 'test-index'}], 'total': 1, 'max_score': 0.19178301}})
+                {'hits': {'hits': [{'_score': ANY, '_type': 'test-type', '_id': '3', '_source': {'age': 16, 'name': 'Joe Justin'}, '_index': 'test-index'}], 'total': 1, 'max_score': ANY}})
 
     def test_mlt_fields(self):
         self.conn.index('test-index', 'test-type', {'name': 'Angus', 'sport': 'football'}, id=3)
@@ -434,7 +428,7 @@ class SearchTestCase(ElasticSearchTestCase):
 
         result = self.conn.more_like_this('test-index', 'test-type', 3, ['sport'], min_term_freq=1, min_doc_freq=1)
         self.assert_result_contains(result,
-                {u'hits': {u'hits': [{u'_score': 0.30685282, u'_type': u'test-type', u'_id': u'4', u'_source': {u'sport': u'football', u'name': u'Cam'}, u'_index': u'test-index'}], u'total': 1, u'max_score': 0.30685282}})
+                {u'hits': {u'hits': [{u'_score': ANY, u'_type': u'test-type', u'_id': u'4', u'_source': {u'sport': u'football', u'name': u'Cam'}, u'_index': u'test-index'}], u'total': 1, u'max_score': ANY}})
 
 
 class DangerousOperationTests(ElasticSearchTestCase):